Fixes #20759: Factor out trait for api test in yaml
